add lagi

+++ application/views/initiative/detail.php
@@ -2,16 +2,15 @@
 	<div class="no_pad" style="margin-bottom: 10px; float:right">
 		<h5 style="color:grey">Initiative</h5>
 		<h4 style="color:grey">
-		<button class="btn btn-success btn-xs" disabled><span class="glyphicon glyphicon-ok"></span></button> 
-		<?php echo $initiative->code?> <?php echo $initiative->title?></h4>
-		<!--<div>
-			<table class="table table-bordered" style="width:250px">
-				<tbody>
-					<tr><td>Start Date</td><td><?php echo $initiative->start?></td></tr>
-					<tr><td>End Date</td><td><?php echo $initiative->end?></td></tr>
-				</tbody>
-			</table>
-		</div><br>-->
+			<?php 
+				if($initiative['stat']=="Delay"){$clr="danger"; $icn="remove";}
+				elseif($initiative['stat']=="In Progress"){$clr="warning"; $icn="refresh";}
+				elseif($initiative['stat']=="Completed"){$clr="success"; $icn="ok";}
+				else{$clr="inverse"; $icn="off";}
+			?>
+			<button class="btn btn-<?php echo $clr?> btn-xs" disabled><span class="glyphicon glyphicon-<?php echo $icn?>"></span></button>
+			<?php echo $initiative['int']->code?> <?php echo $initiative['int']->title?>
+		</h4>
 	</div><div style="clear:both"></div>
 	<h3>Workblock Summary</h3>
 	<div>
@@ -22,7 +21,7 @@
 			<hr>
 			<h3>New Workblock</h3>
 			<form class="form-horizontal" method="post" id="form_src_rm" action="<?php echo base_url();?>workblock/submit_workblock">
-				<input type="hidden" value="<?php echo $initiative->id?>" name="initiative">
+				<input type="hidden" value="<?php echo $initiative['int']->id?>" name="initiative">
 				<div class="form-group">
 					<label class="col-sm-2 control-label">Workblock</label>
 					<div class="col-sm-4">
@@ -59,20 +58,25 @@
 		</div>
 		<table class="table table-bordered" style="margin-top:20px">
 			<thead>
-				<tr class="headertab"><th></th><th>Workblock</th><th>Milestone</th><th>Status</th><th>Start Date</th><th>End Date</th><th style="width:80px"></th></tr>
+				<tr class="headertab"><th></th><th style="width:90px">Workblock</th><th>Milestone</th><th>Start Date</th><th>End Date</th></tr>
 			</thead>
 			<tbody>
 				<?php $i=1; foreach($workblocks as $wb){?>
 				<tr id="wbdtl_<?php echo $wb['wb']->id?>">
-					<td rowspan=<?php echo count($wb['ms'])+1?>></td>
+					<td rowspan=<?php echo count($wb['ms'])+1?>><?php 
+						if($wb['stat']=="Delay"){$clr="danger"; $icn="remove";}
+						elseif($wb['stat']=="In Progress"){$clr="warning"; $icn="refresh";}
+						elseif($wb['stat']=="Completed"){$clr="success"; $icn="ok";}
+						else{$clr="inverse"; $icn="off";}
+					?>
+					<center><button class="btn btn-<?php echo $clr?> btn-xs" disabled><span class="glyphicon glyphicon-<?php echo $icn?>"></span></button></center></td>
 					<td colspan=2><?php echo $i;?>. <a href="<?php echo base_url()?>workblock/detail_workblock/<?php echo $wb['wb']->id?>"><?php echo $wb['wb']->title?></a></td>
-					<td><?php echo $wb['stat'];?></td>
-					<td><?php if($wb['wb']->start){echo date("d M y", strtotime($wb['wb']->start));}?></td>
-					<td><?php if($wb['wb']->end){echo date("d M y", strtotime($wb['wb']->end));}?></td>
-					<td>
+					<td><?php if($wb['wb']->start){echo date("j M y", strtotime($wb['wb']->start));}?></td>
+					<td><?php if($wb['wb']->end){echo date("j M y", strtotime($wb['wb']->end));}?></td>
+					<?php if($user['role']=='admin'){?><td>
 						<button class="btn btn-warning  btn-xs" onclick="toggle_visibility('edit_wb_<?php echo $wb['wb']->id?>');"><span class="glyphicon glyphicon-pencil"></span></button>
 						<button class="btn btn-danger btn-xs" onclick="delete_workblock(<?php echo $wb['wb']->id?>)"><span class="glyphicon glyphicon-trash"></span></button>
-					</td>
+					</td><?php }?>
 					
 				</tr>
 				<tr id="edit_wb_<?php echo $wb['wb']->id?>" style="display:none">	
@@ -113,16 +117,19 @@
 							</div>
 							<input type="submit" class="btn btn-success" style="margin-left:170px">
 						</form>
-					<td>
-						
-					</td>
-					
 				</tr>
 				<?php $j=1; foreach($wb['ms'] as $ms){?>
 				<tr class="ms_wb_<?php echo $wb['wb']->id?>">
-					<td></td><td><?php echo $j.". ".$ms->title?></td><td><?php echo $ms->status?></td>
+					<td><?php 
+						if($ms->status=="Delay"){$clr="danger"; $icn="remove";}
+						elseif($ms->status=="In Progress"){$clr="warning"; $icn="refresh";}
+						elseif($ms->status=="Completed"){$clr="success"; $icn="ok";}
+						else{$clr="inverse"; $icn="off";}
+					?>
+					<center><button class="btn btn-<?php echo $clr?> btn-xs" disabled><span class="glyphicon glyphicon-<?php echo $icn?>"></span></button></center></td>
+					<td><?php echo $j.". ".$ms->title?></td>
 					<td><?php if($ms->start){echo date("j M y", strtotime($ms->start));}?></td>
-					<td><?php if($ms->end){echo date("j M y", strtotime($ms->end));}?></td><td></td>
+					<td><?php if($ms->end){echo date("j M y", strtotime($ms->end));}?></td>
 				</tr>
 				<?php $j++;}?>
 				<?php $i++; }?>
